requirements.txt
# Smart Pharma AI Service - Dependencies
# Python 3.10+ recommended

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database ORM
sqlalchemy==2.0.35
# psycopg[binary]==3.1.18  # Uncomment khi cần kết nối PostgreSQL

# Data Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Environment Variables
python-dotenv==1.0.1
